+++ app/src/main/java/com/tajmoti/tulip/ui/player/streams/AndroidStreamsViewModel.kt
@@ -1,4 +1,4 @@
-package com.tajmoti.tulip.ui.player
+package com.tajmoti.tulip.ui.player.streams
 
 import androidx.lifecycle.*
 import com.tajmoti.libtulip.repository.StreamsRepository
@@ -6,6 +6,7 @@ import com.tajmoti.libtulip.service.*
 import com.tajmoti.libtulip.ui.streams.StreamsViewModel
 import com.tajmoti.libtulip.ui.streams.StreamsViewModelImpl
 import com.tajmoti.tulip.ui.DelegatingViewModel
+import com.tajmoti.tulip.ui.player.VideoPlayerActivityArgs
 import dagger.hilt.android.lifecycle.HiltViewModel
 import javax.inject.Inject